+++ Angular/TunRW-app/backend/controller/events.js
@@ -167,7 +167,7 @@ exports.createEvent = (req,res,next) => {
         creator : req.userData.username
       });
 
-    Event.deleteOne().then(result =>{
+    Event.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result =>{
       if(result.n > 0){
         notification.save().then(notResult => {
           res.status(200).json({
