+++ Angular/TunRW-app/backend/controller/partners.js
@@ -131,7 +131,7 @@ exports.createPartner = (req,res,next) => {
         creator : req.userData.username
       });
 
-      Partner.deleteOne().then(result =>{
+      Partner.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result =>{
         if(result.n > 0){
           notification.save().then(notResult => {
             res.status(200).json({
