+++ Angular/TunRW-app/backend/controller/teamMembers.js
@@ -127,7 +127,7 @@ exports.createTeamMember = (req,res,next) => {
         watched : documents,
         creator : req.userData.username
       });
-      TeamMember.deleteOne().then(result =>{
+      TeamMember.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result =>{
         if(result.n > 0){
           notification.save().then(notResult => {
             res.status(200).json({
